+++ src/main/kotlin/org/zerock/todolist/domain/todo/service/TodoServiceImpl.kt
@@ -5,6 +5,7 @@ import org.springframework.data.domain.Pageable
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
+import org.zerock.todolist.domain.exception.CustomAccessDeniedException
 import org.zerock.todolist.domain.exception.ModelNotFoundException
 import org.zerock.todolist.domain.todo.dto.CreateTodoRequest
 import org.zerock.todolist.domain.todo.dto.TodoResponse
@@ -12,10 +13,13 @@ import org.zerock.todolist.domain.todo.dto.UpdateTodoRequest
 import org.zerock.todolist.domain.todo.model.Todo
 import org.zerock.todolist.domain.todo.model.toResponse
 import org.zerock.todolist.domain.todo.repository.TodoRepository
+import org.zerock.todolist.domain.user.model.User
+import org.zerock.todolist.domain.user.repository.UserRepository
 
 @Service
 class TodoServiceImpl(
-    private val todoRepository: TodoRepository
+    private val todoRepository: TodoRepository,
+    private val userRepository: UserRepository
 ) : TodoService {
 
     override fun getAllTodoList(pageable: Pageable, writer: String?): Page<TodoResponse> {
@@ -32,19 +36,27 @@ class TodoServiceImpl(
     }
 
     @Transactional
-    override fun createTodo(request: CreateTodoRequest): TodoResponse {
+    override fun createTodo(request: CreateTodoRequest, userEmail: String?): TodoResponse {
+        val user = userEmail?.let { userRepository.findByEmail(it) } ?: throw ModelNotFoundException("User", null)
+        println("createTodo - user =====> $user")
         return todoRepository.save(
             Todo(
                 title = request.title,
                 content = request.content,
-                writer = request.writer
+                writer = request.writer,
+                user = user
             )
         ).toResponse()
     }
 
     @Transactional
-    override fun updateTodo(todoId: Long, request: UpdateTodoRequest): TodoResponse {
+    override fun updateTodo(todoId: Long, request: UpdateTodoRequest, userEmail: String?): TodoResponse {
+        val user = userEmail?.let { userRepository.findByEmail(it) } ?: throw ModelNotFoundException("User", null)
         val todo = todoRepository.findByIdOrNull(todoId) ?: throw ModelNotFoundException("Todo", todoId)
+
+        if (todo.user != user) {
+            throw CustomAccessDeniedException("You do not have access.")
+        }
         val (title, content, writer) = request
 
         todo.title = title
@@ -55,26 +67,14 @@ class TodoServiceImpl(
     }
 
     @Transactional
-    override fun deleteTodo(todoId: Long) {
+    override fun deleteTodo(todoId: Long, userEmail: String?) {
+        val user = userEmail?.let { userRepository.findByEmail(it) } ?: throw ModelNotFoundException("User", null)
         val todo = todoRepository.findByIdOrNull(todoId) ?: throw ModelNotFoundException("Todo", todoId)
+
+        if (todo.user != user) {
+            throw CustomAccessDeniedException("You do not have access.")
+        }
+
         todoRepository.delete(todo)
     }
-
-//    override fun paging(pageable: Pageable): Page<Todo> {
-//        val todo = QTodo.todo
-//
-//        val query = from(todo)
-//
-//        query.where(todo.id.gt(pageable.offset))
-////        query.orderBy(pageable.) // TODO: 정렬 기준을 받아 오름차순, 내림차순 정렬
-//        query.limit(pageable.pageSize.toLong())
-//
-//        this.querydsl?.applyPagination(pageable, query)
-//
-//        val content = query.fetch()
-//
-//        val totalCount = query.fetchCount()
-//
-//        return PageImpl(content, pageable, totalCount)
-//    }
 }
